added gamestate static class to hold player locations across threads

+++ src/Unit/Unit.java
@@ -47,11 +47,11 @@ public class Unit {
         return speed;
     }
     
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
     
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
     
@@ -70,6 +70,5 @@ public class Unit {
     public void setSpeed(double speed) {
         this.speed = speed;
     }
-    
     
 }
